~ Fix 46 issues from sweep + 2 broken tests
~ Permission bitmask checks: ordinal() -> getValue() in 3 sites
~ /banksystem deop now revokes admin instead of granting it
~ DeleteBankAccountAsync gate: || -> && so admin OR MANAGE allows
~ BankUploadBlockEntity settings now applied only after ownership check
~ UpdateBankAccountRequest masks permissions, protects bank owner
+ MANAGE permission check on SetAccountName/Icon for client callers
+ Overflow guard in WithdrawMoneyPacket pre-multiplication
~ BankTerminalBlockDataRequest pattern-instanceof instead of unsafe cast

~ isSlaveServerTrustedAsync sends correct FunctionType
~ GetMinimalDataAsync codec: BankManagerData -> BankData
~ GetBankAccountNrByNameAsync handler returns int (not object)
~ 4 async lookups complete(null) on missing account (no more hangs)
~ bank_create_async uses Integer-aware completion handler
~ removeAllBanksAsync uses RemoveAllBanksAsync (was RemoveEmptyBanks)
~ WithdrawItemsFromBankRequest returns empty map on rejection
~ withdraw/has*/toString*/total-balance Async methods call correct delegates

~ convertToRawAmountStatic pads fractional digits to scale-factor width
~ convertToRealAmountStatic float -> double
~ getNormalizedAmountStatic handles <=0, log10, substring bounds

~ ServerBankManager.removeUser collects-then-deletes (no CME)
~ ServerBankAccount.unsubscribeBankChanges direct remove() (no CME)
~ BankAccountChangeStream.nextPacket marked volatile
~ BankSystemMod.init() synchronized, backend volatile

~ TerminalInventory get/remove* return ItemStack.EMPTY (was null)
~ TerminalInventory.load() returns true on success
~ TerminalInventory getItemCount/getFreeSpace fixes
~ BankTerminalBlock.onRemove same-block guard
~ BankDownloadBlock opens menu on main thread
+ MoneyStockpileBlockEntity.dropItems splits over maxStackSize
~ ItemID registration uses server-side variant in server paths

~ BankSystemDataHandler.save_bank early-return when bankManager null
~ money_remove uses getRealBalance() vs real amount
~ DepositItemsInBankRequest clamps negative amounts
~ getNotRemovableItems uses correct list

~ ArithmeticTests: testConvertToRawAmountSingleDecimal expects 1050 (was 1005)
~ AsyncForwardingTests: codec test uses round-trip instead of identity ==
+ Updated changelog/v1.5.0_ALPHA_3.md with full fix list

+++ common/src/main/java/net/kroia/banksystem/banking/bankmanager/BankManager.java
@@ -92,10 +92,10 @@ public class BankManager implements IBankManager {
     }
     public static double convertToRealAmountStatic(long rawAmount)
     {
-        return (float)rawAmount / (float) BankSystemModSettings.ITEM_FRACTION_SCALE_FACTOR;
+        return (double)rawAmount / (double) BankSystemModSettings.ITEM_FRACTION_SCALE_FACTOR;
     }
     public static double convertToRealAmountStatic(long rawAmount, int itemFractionScaleFactor)
     {
-        return (float)rawAmount / (float)itemFractionScaleFactor;
+        return (double)rawAmount / (double)itemFractionScaleFactor;
     }
 }
